C# Live Project
For this project, I was working with a team developing a full scale MVC Web Application in Microsoft Visual Studio, utilizing Entity Framework code first. I got to experience using the AGILE system as well as Microsoft Azure DevOps to work effectively in a team. Each working day, a stand-up meeting was held, in which each member of the team would explain what user story they were working on, what they had accomplished the previous day, and what they were currently working on, as well as if the team members have any issues or roadblocks that they were facing. There was also a sprint planning meeting at the beginning of the project, as well as a retrospective at the end of each week. I mostly worked on stories for the back end of the application. This two week sprint helped me to develop the skills necessary to succeed in a career as a web developer.